Feat/preprod: add preprod environment  (#13)
* chore a bit: move all the service packages under the "internal/service", add config loader, separate server logic to communicate with the environment

* Add test environment warning to server message responses

Introduced a new middleware to the server which adds a warning to every message when the program is running in a pre-production environment. This functionality helps in identifying and separating test and production instances. Also, created a new type PostTgMsgHandler for better structure and readability of the code. This middleware is only active in pre-production environment to avoid any confusion in a live environment.

* fixes: logging in main(), config loading, message handling

* compact warning message